requirements.txt
python-decouple
gitdb
pygit2
git-lfs
GitPython
optparse-pretty